refactor(ui): replace the last emojis in the public frontend with icons (#445)
Seven emojis were still sitting outside the admin panel, which the
earlier cleanup did not cover:

| file | was |
|---|---|
| `app/journal/page.tsx` | 📖 as the empty cover placeholder |
| `components/Lightbox.tsx` | ❤️ / 🤍 on the favourite button |
| `components/ProofingModal.tsx` | ✓, 🔗, 📋 on the copy buttons |
| `components/AdminDiagnosticBanner.tsx` | 🔄 on the reload link |
| `components/DevToolbar.tsx` | 🎨 and ✕ |

## Where the icons come from

The set moved from `app/admin/components/Icons.tsx` to
`components/Icons.tsx`, so both trees use one source. It had no
admin-specific dependency, and the alternative was a second copy of the
same SVG paths in the public components.
`app/admin/components/Icons.tsx` is now a re-export, so not a single
import in the admin panel changed.

## Two CSS adjustments that follow from the swap

- `.infoToggle` (lightbox) is `inline-flex` with a gap now. An emoji is
a glyph in the text flow and needs no alignment of its own; an icon
beside a label does.
- `.journal-card__cover-empty` loses its `font-size`, which only ever
sized the emoji.

## Verification

Browser-checked on an album: the lightbox favourite button renders the
heart icon beside its label and stays aligned; the button reports one
`<svg>` and no emoji in its text content. A repository-wide scan for
emoji outside `app/admin` returns zero hits. 409 unit tests pass, `npm
run build` and `npx tsc --noEmit` clean.
